# contracts/openapi.json
{
  "components": {
    "schemas": {
      "AnonymousPreview": {
        "additionalProperties": false,
        "properties": {
          "alias": {
            "title": "Alias",
            "type": "string"
          },
          "deal_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DealType"
              },
              {
                "type": "null"
              }
            ]
          },
          "employee_band": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employee Band"
          },
          "layer": {
            "default": 1,
            "title": "Layer",
            "type": "integer"
          },
          "notice": {
            "default": "B\u1ea3n xem tr\u01b0\u1edbc n\u1ed9i b\u1ed9; c\u1ea7n ng\u01b0\u1eddi duy\u1ec7t nguy c\u01a1 nh\u1eadn di\u1ec7n tr\u01b0\u1edbc khi chia s\u1ebb.",
            "title": "Notice",
            "type": "string"
          },
          "region": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Region"
              },
              {
                "type": "null"
              }
            ]
          },
          "revenue_band_vnd": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revenue Band Vnd"
          },
          "sector": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Sector"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "alias",
          "sector",
          "region",
          "revenue_band_vnd",
          "employee_band",
          "deal_type"
        ],
        "title": "AnonymousPreview",
        "type": "object"
      },
      "ApiError": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ApiError",
        "type": "object"
      },
...
      "CompanyInput-Input": {
        "additionalProperties": false,
        "properties": {
          "address": {
            "$ref": "#/components/schemas/Fact_Annotated_str__FieldInfo_annotation_NoneType__required_True__metadata__MinLen_min_length_1___MaxLen_max_length_2000_____"
          },
          "company_name": {
            "$ref": "#/components/schemas/Fact_Annotated_str__FieldInfo_annotation_NoneType__required_True__metadata__MinLen_min_length_1___MaxLen_max_length_200_____"
          },
          "customer_groups": {
            "$ref": "#/components/schemas/Fact_Annotated_str__FieldInfo_annotation_NoneType__required_True__metadata__MinLen_min_length_1___MaxLen_max_length_2000_____"
          },
          "deal_type": {
            "$ref": "#/components/schemas/Fact_DealType_"
          },
          "decision_maker": {
            "$ref": "#/components/schemas/Fact_Annotated_str__FieldInfo_annotation_NoneType__required_True__metadata__MinLen_min_length_1___MaxLen_max_length_2000_____"
          },
          "description": {
            "$ref": "#/components/schemas/Fact_Annotated_str__FieldInfo_annotation_NoneType__required_True__metadata__MinLen_min_length_1___MaxLen_max_length_2000_____"
          },
          "employees": {
            "$ref": "#/components/schemas/Fact_Annotated_int__FieldInfo_annotation_NoneType__required_True__metadata__Ge_ge_0___Le_le_10000000_____"
          },
          "financials": {
            "items": {
              "$ref": "#/components/schemas/FinancialYear-Input"
            },
            "maxItems": 3,
            "title": "Financials",
            "type": "array"
          },
          "founded_year": {
            "$ref": "#/components/schemas/Fact_Annotated_int__FieldInfo_annotation_NoneType__required_True__metadata__Ge_ge_1800___Le_le_2100_____"
          },
          "funds_destination": {
            "$ref": "#/components/schemas/Fact_Annotated_str__FieldInfo_annotation_NoneType__required_True__metadata__MinLen_min_length_1___MaxLen_max_length_2000_____"
          },
          "objectives": {
            "$ref": "#/components/schemas/Fact_Annotated_str__FieldInfo_annotation_NoneType__required_True__metadata__MinLen_min_length_1___MaxLen_max_length_2000_____"
          },
          "permitted_use": {
            "$ref": "#/components/schemas/Fact_Annotated_str__FieldInfo_annotation_NoneType__required_True__metadata__MinLen_min_length_1___MaxLen_max_length_2000_____"
          },
          "region": {
            "$ref": "#/components/schemas/Fact_Region_"
          },
          "sector": {
            "$ref": "#/components/schemas/Fact_Sector_"
          },
          "shareholders": {
            "$ref": "#/components/schemas/Fact_Annotated_str__FieldInfo_annotation_NoneType__required_True__metadata__MinLen_min_length_1___MaxLen_max_length_2000_____"
          },
          "stake_percent": {
            "$ref": "#/components/schemas/Fact_Annotated_Decimal__FieldInfo_annotation_NoneType__required_True__metadata__Ge_ge_0___Le_le_100____PydanticGeneralMetadata_decimal_places_2_____-Input"
          },
          "tax_id": {
            "$ref": "#/components/schemas/Fact_Annotated_str__FieldInfo_annotation_NoneType__required_True__metadata__MinLen_min_length_3___MaxLen_max_length_30_____"
          },
          "technology": {
            "$ref": "#/components/schemas/Fact_Annotated_str__FieldInfo_annotation_NoneType__required_True__metadata__MinLen_min_length_1___MaxLen_max_length_2000_____"
          }
        },
        "required": [
          "company_name"
        ],
        "title": "CompanyInput",
        "type": "object"
      },
...
      "CompanyResponse": {
        "additionalProperties": false,
        "properties": {
          "alias": {
            "title": "Alias",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "profile": {
            "$ref": "#/components/schemas/CompanyInput-Output"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "alias",
          "revision",
          "created_at",
          "updated_at",
          "profile"
        ],
        "title": "CompanyResponse",
        "type": "object"
      },
      "CompanySummary": {
        "additionalProperties": false,
        "properties": {
          "alias": {
            "title": "Alias",
            "type": "string"
          },
          "deal_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DealType"
              },
              {
                "type": "null"
              }
            ]
          },
          "employee_band": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employee Band"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "region": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Region"
              },
              {
                "type": "null"
              }
            ]
          },
          "revenue_band": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revenue Band"
          },
          "sector": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Sector"
              },
              {
                "type": "null"
              }
            ]
          },
          "stake_percent": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stake Percent"
          },
          "tax_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tax Id"
          },
          "total_facts_count": {
            "default": 0,
            "title": "Total Facts Count",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "verified_facts_count": {
            "default": 0,
            "title": "Verified Facts Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "alias",
          "name",
          "tax_id",
          "updated_at"
        ],
        "title": "CompanySummary",
        "type": "object"
      },
      "CompanyUpdate": {
        "additionalProperties": false,
        "properties": {
          "expected_revision": {
            "minimum": 1.0,
            "title": "Expected Revision",
            "type": "integer"
          },
          "profile": {
            "$ref": "#/components/schemas/CompanyInput-Input"
          }
        },
        "required": [
          "expected_revision",
          "profile"
        ],
        "title": "CompanyUpdate",
        "type": "object"
      },
      "DealType": {
        "enum": [
          "PRIMARY",
          "SECONDARY",
          "MIXED"
        ],
        "title": "DealType",
        "type": "string"
      },
...
      "Fact_Annotated_str__FieldInfo_annotation_NoneType__required_True__metadata__MinLen_min_length_1___MaxLen_max_length_2000_____": {
        "additionalProperties": false,
        "properties": {
          "entered_by": {
            "default": "",
            "maxLength": 120,
            "title": "Entered By",
            "type": "string"
          },
          "issue": {
            "$ref": "#/components/schemas/Issue",
            "default": "NONE"
          },
          "notes": {
            "default": "",
            "maxLength": 1000,
            "title": "Notes",
            "type": "string"
          },
          "reviewed_by": {
            "default": "",
            "maxLength": 120,
            "title": "Reviewed By",
            "type": "string"
          },
          "reviewed_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed On"
          },
          "source": {
            "default": "",
            "maxLength": 500,
            "title": "Source",
            "type": "string"
          },
          "source_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Date"
          },
          "value": {
            "anyOf": [
              {
                "maxLength": 2000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "verification": {
            "$ref": "#/components/schemas/Verification",
            "default": "SELF_DECLARED"
          }
        },
        "title": "Fact[Annotated[str, FieldInfo(annotation=NoneType, required=True, metadata=[MinLen(min_length=1), MaxLen(max_length=2000)])]]",
        "type": "object"
      },
...
      "FinancialYear-Input": {
        "additionalProperties": false,
        "properties": {
          "ebitda_vnd": {
            "$ref": "#/components/schemas/Fact_Annotated_Decimal__FieldInfo_annotation_NoneType__required_True__metadata___PydanticGeneralMetadata_max_digits_18__decimal_places_2_____-Input"
          },
          "revenue_vnd": {
            "$ref": "#/components/schemas/Fact_Annotated_Decimal__FieldInfo_annotation_NoneType__required_True__metadata__Ge_ge_0____PydanticGeneralMetadata_max_digits_18__decimal_places_2_____-Input"
          },
          "year": {
            "maximum": 2100.0,
            "minimum": 1900.0,
            "title": "Year",
            "type": "integer"
          }
        },
        "required": [
          "year"
        ],
        "title": "FinancialYear",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IdentifiedPreview": {
        "additionalProperties": false,
        "properties": {
          "facts": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Fact_Union_Annotated_str__FieldInfo_annotation_NoneType__required_True__metadata__MinLen_min_length_1___MaxLen_max_length_2000______Decimal__int__"
            },
            "title": "Facts",
            "type": "object"
          },
          "financials": {
            "items": {
              "$ref": "#/components/schemas/FinancialYear-Output"
            },
            "title": "Financials",
            "type": "array"
          },
          "layer": {
            "default": 2,
            "title": "Layer",
            "type": "integer"
          },
          "notice": {
            "default": "B\u1ea3n xem tr\u01b0\u1edbc n\u1ed9i b\u1ed9; kh\u00f4ng ph\u1ea3i quy\u1ec1n chia s\u1ebb cho buyer.",
            "title": "Notice",
            "type": "string"
          }
        },
        "required": [
          "facts",
          "financials"
        ],
        "title": "IdentifiedPreview",
        "type": "object"
      },
      "Issue": {
        "enum": [
          "NONE",
          "CONFLICTED",
          "EXPIRED"
        ],
        "title": "Issue",
        "type": "string"
      },
      "Region": {
        "enum": [
          "NORTH",
          "CENTRAL",
          "SOUTH"
        ],
        "title": "Region",
        "type": "string"
      },
      "RestrictedPreview": {
        "additionalProperties": false,
        "properties": {
          "facts": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Fact_Annotated_str__FieldInfo_annotation_NoneType__required_True__metadata__MinLen_min_length_1___MaxLen_max_length_2000_____"
            },
            "title": "Facts",
            "type": "object"
          },
          "layer": {
            "default": 3,
            "title": "Layer",
            "type": "integer"
          },
          "notice": {
            "default": "N\u1ed9i dung h\u1ea1n ch\u1ebf v\u00e0 ph\u00f2ng d\u1eef li\u1ec7u; ch\u01b0a tri\u1ec3n khai NDA/ph\u00ea duy\u1ec7t hay quy\u1ec1n buyer.",
            "title": "Notice",
            "type": "string"
          }
        },
        "required": [
          "facts"
        ],
        "title": "RestrictedPreview",
        "type": "object"
      },
...
      "Sector": {
        "enum": [
          "SOFTWARE",
          "MANUFACTURING",
          "SERVICES",
          "OTHER"
        ],
        "title": "Sector",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "Verification": {
        "enum": [
          "SELF_DECLARED",
          "DOCUMENT_VERIFIED",
          "THIRD_PARTY_CONFIRMED"
        ],
        "title": "Verification",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "VAULT internal profiles and local synthetic data rooms. No buyer disclosure or authentication is implemented.",
    "title": "VAULT Core",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/companies": {
      "get": {
        "operationId": "list_companies",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 200,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sector",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Sector"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sector"
            }
          },
          {
            "in": "query",
            "name": "region",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Region"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region"
            }
          },
          {
            "in": "query",
            "name": "deal_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/DealType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Deal Type"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CompanySummary"
                  },
                  "title": "Response List Companies",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "List Companies",
        "tags": [
          "Internal company drafts"
        ]
      },
      "post": {
        "operationId": "create_company",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompanyInput-Input"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Create Company",
        "tags": [
          "Internal company drafts"
        ]
      }
    },
    "/api/v1/companies/{company_id}": {
      "get": {
        "operationId": "get_company",
        "parameters": [
          {
            "in": "path",
            "name": "company_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Company Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Company",
        "tags": [
          "Internal company drafts"
        ]
      },
      "put": {
        "operationId": "update_company",
        "parameters": [
          {
            "in": "path",
            "name": "company_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Company Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompanyUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Update Company",
        "tags": [
          "Internal company drafts"
        ]
      }
    },
    "/api/v1/companies/{company_id}/preview/{layer}": {
      "get": {
        "operationId": "preview_company",
        "parameters": [
          {
            "in": "path",
            "name": "company_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Company Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "layer",
            "required": true,
            "schema": {
              "maximum": 3,
              "minimum": 1,
              "title": "Layer",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AnonymousPreview"
                    },
                    {
                      "$ref": "#/components/schemas/IdentifiedPreview"
                    },
                    {
                      "$ref": "#/components/schemas/RestrictedPreview"
                    }
                  ],
                  "title": "Response Preview Company"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Preview Company",
        "tags": [
          "Internal company drafts"
        ]
      }
    },
...
  }
}